Rename files to React components style
