+++ src/components/Logo/LogoSection.tsx
@@ -7,7 +7,7 @@ import { useDispatch, useSelector } from 'react-redux';
 // project import
 import { Logo } from './Logo';
 import { activeItem } from '../../store/reducers/menuSlice';
-import { RootState } from '../../store';
+import { RootState } from '../../store/reduxStore';
 
 // ==============================|| MAIN LOGO ||============================== //
