+++ src/index.tsx
@@ -2,10 +2,10 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { HashRouter } from 'react-router-dom';
 import { Provider as ReduxProvider } from 'react-redux';
-import { store } from './store';
+import { reduxStore } from './store/reduxStore';
 import { ScrollTop } from './components/ScrollTop';
 import { ThemeCustomization } from './themes/ThemeCustomization';
-import { Routes } from './routes';
+import { Routes } from './Routes';
 
 const App = () => {
   return (
@@ -21,7 +21,7 @@ const root = createRoot(document.getElementById('root')!);
 
 root.render(
   <StrictMode>
-    <ReduxProvider store={store}>
+    <ReduxProvider store={reduxStore}>
       <HashRouter>
         <App />
       </HashRouter>
