+++ src/utils/passwordStrength.ts
